+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2021 - As admin or organizer, delete all meetings if agenda not published
 
 ## [2.12.1] - 2019-09-25
 ### Fixed